Changed the approach on deleting old posts
Previous implementation could delete a post when processing a rank
position that had a post that now has a lower rating and making
the algorithm having to re insert it and resend it to monkeylearn.

On the other hand since we are not using transactional writes
well, we could mess things up but previous implementation didn't
handle it very well eighter.